refactor: change pages to be in a folder
