+++ src/App.js
@@ -1,27 +1,27 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { useEffect, useState } from 'react';
 
-import LoginScreen from "./pages/LoginScreen";
-import SignupScreen from "./pages/SignupScreen";
-import HomeScreen from "./pages/Home/HomeScreen";
-import ProductScreen from "./pages/ProductScreen";
-import CheckoutScreen from "./pages/CheckoutScreen";
-import CartScreen from "./pages/CartScreen";
+import LoginScreen from './pages/LoginScreen';
+import SignupScreen from './pages/SignupScreen';
+import HomeScreen from './pages/Home/HomeScreen';
+import ProductScreen from './pages/ProductScreen';
+import CheckoutScreen from './pages/CheckoutScreen';
+import CartScreen from './pages/Cart/CartScreen';
 
-import UserContext from "./contexts/UserContext";
+import UserContext from './contexts/UserContext';
 
-import Header from "./shared/Header";
+import Header from './shared/Header';
 
 function App() {
-  const [userToken, setUserToken] = useState(localStorage.getItem("token"));
+  const [userToken, setUserToken] = useState(localStorage.getItem('token'));
   useEffect(() => {
     if (userToken) {
-      localStorage.setItem("token", userToken);
+      localStorage.setItem('token', userToken);
     } else {
-      localStorage.setItem("token", "");
+      localStorage.setItem('token', '');
     }
   }, [userToken]);
-  const URL = "https://xablaum.herokuapp.com";
+  const URL = 'https://xablaum.herokuapp.com';
 
   return (
     <UserContext.Provider
